requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
html2text==2020.1.16
